test: add unit tests for openGauss vector store component
